pyproject.toml
[project]
name = "egile-agent-hub"
version = "0.1.0"
description = "Multi-agent hub for running multiple Egile agents and teams in a unified AgentOS"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "egile-agent-core @ file:///C:/Users/jeanb/OneDrive/Documents/projects/egile-agent-core",
    "egile-agent-prospectfinder @ file:///C:/Users/jeanb/OneDrive/Documents/projects/egile-agent-prospectfinder",
    "egile-agent-x-twitter @ file:///C:/Users/jeanb/OneDrive/Documents/projects/egile-agent-x-twitter",
    "agno>=2.3.0",
    "httpx>=0.27.0",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
all = [
    "egile-agent-core @ file:///C:/Users/jeanb/OneDrive/Documents/projects/egile-agent-core",
    "egile-agent-prospectfinder @ file:///C:/Users/jeanb/OneDrive/Documents/projects/egile-agent-prospectfinder",
    "egile-agent-x-twitter @ file:///C:/Users/jeanb/OneDrive/Documents/projects/egile-agent-x-twitter",
    "agno>=2.3.0",
    "httpx>=0.27.0",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
agent-hub = "egile_agent_hub.run_server:main"
agent-hub-agentos = "egile_agent_hub.run_server:run_agentos_only"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/egile_agent_hub"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
